+++ src/react/calculator/src/App.js
@@ -1,147 +1,38 @@
+import React from 'react';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import './App.css';
-import MultiSelectDropdown from "./MultiSelectDropdown";
-import { BrowserRouter, Route, Link } from "react-router-dom";
-import { v4 as uuid } from 'uuid';
-import { appendErrors, useForm } from "react-hook-form";
-//import Select from 'react-select'
+import { useForm } from "react-hook-form";
 import useFetch from "react-fetch-hook";
 
 
 
-const contributor = "Luca"
-/*
-  < div class="row" >
-  {
-    mates.map(mate => <label>
-      <p>{mate.firstname}</p>
-      <p>{mate.lastname}</p>
-      <p>{mate.personID}</p>
-      <p>{mate.groupID}</p>
-    </label>)
-  }
-        </div >
-*/
 function App() {
-  const { isLoading, data } = useFetch("https://8080-nklsdhbw-webprogramming-ltpyo05qis6.ws-eu81.gitpod.io/api/bills");
-  const { register, handleSubmit, formState: { errors } } = useForm();
+    let { isLoading, data } = useFetch("https://8080-nklsdhbw-webprogramming-ltpyo05qis6.ws-eu81.gitpod.io/api/login");
 
+    const onSubmit = formData => {
+        console.log(data)
+        console.log(formData)
+        data.forEach(element => {
+            if ((formData.username === element.username) && (formData.password === element.password)) {
+                console.log("perfekt")
+            }
+        });
+    }
+    const { register, handleSubmit, formState: { errors } } = useForm();
 
-  if (isLoading) {
-    console.log("...loading")
-    return <div>Is loading!</div>
-  }
-
-  const bills = data;
-
-
-  const onSubmit = data => {
-    let date = new Date();
-    date = date.toISOString()
-    date = date.substring(0, 10)
-
-
-    let datasharedWith = data.sharedWith
-    let amountPeople = datasharedWith.length + 1
-    datasharedWith.forEach(element => {
-
-
-      fetch("https://8080-nklsdhbw-webprogramming-ltpyo05qis6.ws-eu81.gitpod.io/api/bills?contributor=" + contributor + "&amount=" + parseInt((data.amount / amountPeople)) + "&sharedWith=" + element + "&comment=" + data.comment + "&billID=" + uuid() + "&date=" + date, {
-
-        headers: {
-          'Accept': 'application/json',
-          'Content-Type': 'application/json'
-        },
-        method: "PUT"
-      })
-        .then(function (res) { window.location.reload() })
-        .catch(function (res) { console.log(res) })
-    });
-  }
-
-  const options = [
-    { value: 'luca', label: 'Luca' },
-    { value: 'niklas', label: 'Niklas' },
-    { value: 'simon', label: 'Simon' }
-  ]
-  const MyComponent = () => (
-    <Select options={options} isMulti />
-  )
-
-
-
-
-
-
-
-
-
-  return (
-
-    <main class="form-signin w-100 m-auto">
-
-      <form onSubmit={handleSubmit(onSubmit)}>
-        <div class="row">
-          <div class="col">
-            <div class="input-group mb-3">
-              <span class="input-group-text">Betrag €</span>
-              <input {...register("amount")} type="text" class="form-control" aria-label="Amount (to the nearest dollar)"></input>
-
+    return (
+        <form onSubmit={handleSubmit(onSubmit)}>
+            <div className="form-group">
+                <label htmlFor="email">Email address</label>
+                <input {...register("username")} type="email" className="form-control" id="email" aria-describedby="emailHelp" />
+                <small id="emailHelp" className="form-text text-muted">We'll never share your email with anyone else.</small>
             </div>
-
-          </div>
-          <div class="col checkbox">
-            <div>
-              <input {...register("sharedWith")} type="checkbox" id="option1" value="niklas" />
-              <label for="option1">Niklas</label>
+            <div className="form-group">
+                <label htmlFor="password">Password</label>
+                <input {...register("password")} type="password" className="form-control" id="password" />
             </div>
-            <div>
-              <input {...register("sharedWith")} type="checkbox" id="option2" value="luca" />
-              <label for="option2">Luca</label>
-            </div>
-            <div>
-              <input {...register("sharedWith")} type="checkbox" id="option3" value="tim" />
-              <label for="option3">Tim</label>
-            </div>
-            <div>
-              <input {...register("sharedWith")} type="checkbox" id="option4" value="tom" />
-              <label for="option4">Tom</label>
-            </div>
-
-
-
-
-          </div>
-
-        </div>
-        <div class="row">
-          <div class="col">
-            <div class="input-group " id="Kommentar">
-              <span class="input-group-text">Kommentar</span>
-              <textarea {...register("comment")} class="form-control" aria-label="Kommentar"></textarea>
-            </div>
-          </div>
-        </div>
-        <div class="row">
-          <div class="col">
-            <div>
-              <button id="create" class="w-100 btn btn-lg btn-primary" type="submit">
-                splitten
-              </button>
-            </div>
-          </div>
-        </div>
-        <div class="row">
-          <p class="text-start">Placeholder</p>
-        </div>
-
-
-      </form>
-    </main >
-
-  );
-
-
+            <button type="submit" className="btn btn-primary">Submit</button>
+        </form>
+    );
 }
 
 export default App;
